# workspace/fairsplit/api-spec.yaml
openapi: 3.1.0
info:
  title: FairSplit API
  version: 1.0.0
  description: |
    Zero-dependency shared expense tracker API.

    ## Authentication
    All group-scoped endpoints require a valid member session token.
    The token is issued by `POST /api/v1/groups` (group creation) or
    `POST /api/v1/groups/{group_id}/members` (join via invite link).

    **Primary:** `httpOnly` cookie named `fairsplit_member_token`.
    **Fallback:** `Authorization: Bearer <token>` header (API clients).

    The token is a JWT signed with HS256 containing:
    ```json
    {
      "sub": "<member_id>",
      "group_id": "<group_id>",
      "is_admin": true,
      "iat": 1713200000,
      "exp": 1715792000
    }
    ```

    ## Response Envelopes
    **Singleton responses** (auth, single-resource GET/POST/PATCH/DELETE):
    The resource object is returned directly at the top level. No wrapper.

    **List responses** (collection GET endpoints):
    ```json
    {
      "data": [...],
      "total": 23,
      "page": 1,
      "per_page": 20
    }
    ```

    **Error responses** (all 4xx and 5xx):
    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Human-readable description.",
        "field": "amount",
        "request_id": "req_abc123"
      }
    }
    ```

    ## Idempotency
    Mutation endpoints (`POST` for create, `DELETE` for soft-delete) accept an
    `X-Idempotency-Key` header containing a client-generated UUID. The server
...

  contact:
...
servers:
...
paths:
...
  /groups/{group_id}/expenses:
    get:
      operationId: listExpenses
      summary: List expenses in a group (paginated)
      description: |
        Returns active (not soft-deleted) expenses for a group, sorted by
        `created_at DESC`. Includes payer name and split count per expense.
        For full split breakdown, use `GET /groups/{id}/expenses/{expense_id}`.

        **Pagination:** offset-based. Default: page=1, per_page=20. Maximum per_page=100.
      tags: [expenses]
      security:
        - MemberCookie: []
        - BearerToken: []
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated expense list
          content:
            application/json:
              schema:
                type: object
                required: [data, total, page, per_page]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ExpenseSummary"
                  total:
                    type: integer
                    description: Total count of active expenses in the group
                    example: 23
                  page:
                    type: integer
                    example: 1
                  per_page:
                    type: integer
                    example: 20
              example:
                data:
                  - id: "c5dd3013-3h6g-6g9c-0e6h-8f6c3g4d5e6f"
                    description: "Airbnb deposit"
                    amount_cents: 84000
                    amount_display: "$840.00"
                    currency: USD
                    split_type: equal
                    split_count: 6
                    payer_id: "a3bb1891-1f4e-4e7a-8c4f-6d4a1e2b3c4d"
                    payer_name: "Maya"
                    logged_by_name: "Maya"
                    expense_date: "2026-04-15"
                    created_at: "2026-04-15T14:10:00Z"
                total: 23
                page: 1
                per_page: 20
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      operationId: createExpense
      summary: Log a new expense
      description: |
        Creates an expense with a full split breakdown. The split must satisfy:
        - `equal`: amounts are computed server-side; `splits` array must list
          all member IDs to include in the split (amounts are ignored if provided).
        - `custom_amount`: `splits[].amount_cents` must sum exactly to
          `amount_cents`. Returns 422 `SPLIT_SUM_MISMATCH` otherwise.
        - `custom_percentage`: `splits[].percentage` must sum exactly to 100.0.
          Server converts to cents; any 1-cent rounding remainder is added to
          the first member's split, alphabetically.

        `payer_member_id` does not need to be in the `splits` array (payer
        may be paying for non-participants — allowed by design).

        **Idempotent:** provide `X-Idempotency-Key`. Double-submits return the
        original 201 without creating a duplicate expense.

        **Rate limit:** 60 requests per member per hour.
      tags: [expenses]
      security:
        - MemberCookie: []
        - BearerToken: []
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
            examples:
              equal_split:
                summary: Equal split among all members
                value:
                  description: "Airbnb deposit"
                  amount_cents: 84000
                  currency: "USD"
                  payer_member_id: "a3bb1891-1f4e-4e7a-8c4f-6d4a1e2b3c4d"
                  split_type: "equal"
                  expense_date: "2026-04-15"
                  splits:
                    - member_id: "a3bb1891-1f4e-4e7a-8c4f-6d4a1e2b3c4d"
                    - member_id: "b4cc2902-2g5f-5f8b-9d5g-7e5b2f3c4d5e"
                    - member_id: "c5dd3013-3h6g-6g9c-0e6h-8f6c3g4d5e6f"
              custom_amount:
                summary: Custom amount split
                value:
                  description: "Groceries"
                  amount_cents: 4550
                  currency: "USD"
                  payer_member_id: "a3bb1891-1f4e-4e7a-8c4f-6d4a1e2b3c4d"
                  split_type: "custom_amount"
                  expense_date: "2026-04-15"
                  splits:
                    - member_id: "a3bb1891-1f4e-4e7a-8c4f-6d4a1e2b3c4d"
                      amount_cents: 2275
                    - member_id: "b4cc2902-2g5f-5f8b-9d5g-7e5b2f3c4d5e"
                      amount_cents: 2275
              custom_percentage:
                summary: Custom percentage split
                value:
                  description: "Shared Uber"
                  amount_cents: 2400
                  currency: "USD"
                  payer_member_id: "b4cc2902-2g5f-5f8b-9d5g-7e5b2f3c4d5e"
                  split_type: "custom_percentage"
                  expense_date: "2026-04-15"
                  splits:
                    - member_id: "a3bb1891-1f4e-4e7a-8c4f-6d4a1e2b3c4d"
                      percentage: 60.0
                    - member_id: "b4cc2902-2g5f-5f8b-9d5g-7e5b2f3c4d5e"
                      percentage: 40.0
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "409":
          description: Group is archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: GROUP_ARCHIVED
                  message: "Cannot add expenses to an archived group."
                  request_id: req_abc123
        "422":
          description: Validation error or split sum mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                split_mismatch:
                  summary: Split amounts don't sum to expense total
                  value:
                    error:
                      code: SPLIT_SUM_MISMATCH
                      message: "Split amounts sum to 4500 cents but expense total is 4550 cents."
                      request_id: req_abc123
                validation_error:
                  summary: Missing required field
                  value:
                    error:
                      code: VALIDATION_ERROR
                      message: "Description is required."
                      field: description
                      request_id: req_abc123
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    MemberCookie:
      type: apiKey
      in: cookie
      name: fairsplit_member_token
      description: |
        httpOnly session cookie issued on group creation or group join.
        Set by the server. Cannot be accessed by JavaScript.
        Expiry: 30 days, rolling (reset on each valid request).
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Fallback for API clients that cannot use cookies.
        Same JWT as the cookie value. Pass in Authorization: Bearer <token> header.
  parameters:
    GroupId:
      name: group_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: The group's UUID
...
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: |
        Client-generated UUID for idempotency protection. Generate one UUID
        when the form opens and reuse it on all retries. The server returns
        the original response for duplicate keys (within 24 hours).
  responses:
    Unauthorized:
      description: No session or session expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            no_session:
              value:
                error:
                  code: NO_SESSION
                  message: "No session token found. Please join the group first."
                  request_id: req_abc123
            session_expired:
              value:
                error:
                  code: SESSION_EXPIRED
                  message: "Your session has expired. Please re-enter your name to continue."
                  request_id: req_abc123
    Forbidden:
      description: Valid session but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            permission_denied:
              value:
                error:
                  code: PERMISSION_DENIED
                  message: "You can only edit expenses you logged, or you must be the group admin."
                  request_id: req_abc123
            wrong_group:
              value:
                error:
                  code: MEMBER_WRONG_GROUP
                  message: "Your session token is for a different group."
                  request_id: req_abc123
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: GROUP_NOT_FOUND
              message: "Group not found. The link may be invalid."
              request_id: req_abc123
...
  schemas:
...
    ExpenseSummary:
      type: object
      description: Lightweight expense record for list view (no full splits array)
      required:
        - id
        - description
        - amount_cents
        - amount_display
        - currency
        - split_type
        - split_count
        - payer_id
        - payer_name
        - logged_by_name
        - expense_date
        - created_at
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        amount_cents:
          type: integer
          minimum: 1
        amount_display:
          type: string
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
        split_type:
          type: string
          enum: [equal, custom_amount, custom_percentage]
        split_count:
          type: integer
          description: Number of members in this expense's split
        payer_id:
          type: string
          format: uuid
        payer_name:
          type: string
        logged_by_name:
          type: string
        expense_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - GROUP_NOT_FOUND
                - MEMBER_NOT_FOUND
                - EXPENSE_NOT_FOUND
                - SETTLEMENT_NOT_FOUND
                - INVALID_INVITE_TOKEN
                - GROUP_ARCHIVED
                - GROUP_HAS_OUTSTANDING_BALANCES
                - PERMISSION_DENIED
                - NO_SESSION
                - SESSION_EXPIRED
                - SESSION_INVALID
                - MEMBER_WRONG_GROUP
                - IDEMPOTENCY_CONFLICT
                - SPLIT_SUM_MISMATCH
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error description
            field:
              type: string
              nullable: true
              description: The specific field that failed validation (for VALIDATION_ERROR only)
            request_id:
              type: string
              description: Unique request identifier for log correlation